Fixed performance and de-enhancing bugs
There was a bug where closing the student or match datacard window caused search filters to reset. Was fixed.

Also made it so that all subthreads close now on exit, by setting them to background.

Also made it so that the progress bar actually updates with progress, which was a simple division by two integers always making 0.

Also added some statistics parameters to the home screen, namely the number of students passed and the total number of students, in or out.

Tested the application a little more. Seems stable.

Updated installer.

+++ Reindeer Hunter/Matcher.cs
@@ -106,7 +106,7 @@ namespace Reindeer_Hunter
                         "Forgot to supply matchesDone or matchesToProcess");
 
                 // Calculate how far we've gone.
-                decimalPercent = 100 * (matchesDone / matchesToProcess);
+                decimalPercent = 100 * ((double)matchesDone / matchesToProcess);
                 int percent = (int)Math.Round(decimalPercent);
 
                 // Create the actual string message to be sent
